Fixed a crash with the SparseLightmapTransformManager
The crash occurred when the transformed lightmap data got deleted while the render thread was going to interact with it and accessed garbage memory. New implementation now makes sure to wait for the render thread to complete and then removes it. To mitigate the time waiting for the render thread it removes the transformed lightmap data in batches.